Inherited tables, consistency polished (awaiting full testing)

+++ Fitness_Appv2/Services/SetLogDataModel.cs
@@ -8,8 +8,8 @@ namespace Fitness_Appv2.Services
     public class SetLogDataModel { 
 
         public int Index { get; set; }
-        public int Reps { get; set; }
-        public int Weight{ get; set; }
+        public float Reps { get; set; }
+        public float Weight{ get; set; }
     }
 }
